+++ Runtime/Utility/Utils.cs
@@ -9,6 +9,11 @@ namespace Eflatun.SceneReference.Utility
     internal static class Utils
     {
         /// <summary>
+        /// GUID of all zeros. Invalid assets have this GUID.
+        /// </summary>
+        public const string AllZeroGuid = "00000000000000000000000000000000";
+
+        /// <summary>
         /// Returns the given <paramref name="path"/> without file extension.
         /// </summary>
         public static string WithoutExtension(this string path) => path.BeforeLast('.');
@@ -21,13 +26,24 @@ namespace Eflatun.SceneReference.Utility
         /// <summary>
         /// Returns the portion of <paramref name="source"/> that comes before the last occurence of <paramref name="chr"/>.
         /// </summary>
-        public static string BeforeLast(this string source, char chr) =>
-            source.Substring(0, source.LastIndexOf(chr));
+        public static string BeforeLast(this string source, char chr)
+        {
+            var lastChrIndex = source.LastIndexOf(chr);
+            return lastChrIndex < 0
+                ? source
+                : source.Substring(0, lastChrIndex);
+        }
 
         /// <summary>
-        /// Returns if the given <paramref name="guidHex"/> is valid. A valid GUID hex is 32 chars of hexadecimals.
+        /// Returns if the given <paramref name="guid"/> is valid. A valid GUID is 32 chars of hexadecimals.
         /// </summary>
-        public static bool IsValidGuidHex(this string guidHex) =>
-            guidHex.Length == 32 && guidHex.ToUpper().All("0123456789ABCDEF".Contains);
+        public static bool IsValidGuid(this string guid) =>
+            guid.Length == 32 && guid.ToUpper().All("0123456789ABCDEF".Contains);
+
+        /// <summary>
+        /// If the given GUID is null or whitespace returns <see cref="AllZeroGuid"/>. Otherwise returns as-is.
+        /// </summary>
+        public static string GuardGuidAgainstNullOrWhitespace(this string guid) =>
+            string.IsNullOrWhiteSpace(guid) ? AllZeroGuid : guid;
     }
 }